+++ jest.config.js
@@ -1,0 +1,5 @@
+/** @type {import('jest').Config} */
+module.exports = {
+  testEnvironment: "jsdom",
+  testMatch: ["<rootDir>/content/*.test.js"],
+};